packages/embedding-pipeline/requirements.txt
# Core dependencies
chromadb==0.4.24
sentence-transformers==2.5.1
pymysql==1.1.0
sqlalchemy==2.0.25
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
click==8.1.7
schedule==1.2.0

# API client for ChromaDB
httpx==0.25.2
tenacity==8.2.3